Refactor the fitbit single daily values
