+++ src/helpers/daily-data-providers/fitbit-hrv.ts
@@ -1,20 +1,5 @@
-﻿import { add } from "date-fns";
-import queryAllDeviceData from "./query-all-device-data";
+﻿import queryFitbitSingleDailyValue from "./query-fitbit-single-daily-value";
 
 export default function (startDate: Date, endDate: Date) {
-	return queryAllDeviceData({
-		namespace: "Fitbit",
-		type: "Hrv",
-		observedAfter: add(startDate, { days: -1 }).toISOString(),
-		observedBefore: add(endDate, { days: 1 }).toISOString()
-	}).then(function (result) {
-		var data: { [key: string]: number } = {};
-		result.forEach((d) => {
-			if (!d.startDate) { return; }
-			if (parseFloat(d.value) == 0) { return; }
-			var dayKey = d.startDate!.substr(0, 10);
-			data[dayKey] = Math.round(parseFloat(d.value));
-		});
-		return data;
-	});
+	return queryFitbitSingleDailyValue("Hrv", startDate, endDate);
 }
